+++ examples/docs-site/src/app/page.tsx
@@ -1,5 +1,4 @@
 import { HeroSection } from '@/components/HeroSection'
-import { InteractiveDemo } from '@/components/InteractiveDemo'
 import { InstallSection } from '@/components/InstallSection'
 import { PatternSection } from '@/components/PatternSection'
 import { PrimitivesSection } from '@/components/PrimitivesSection'
@@ -10,7 +9,6 @@ export default function Home() {
   return (
     <main>
       <HeroSection />
-      <InteractiveDemo />
       <InstallSection />
       <PatternSection />
       <PrimitivesSection />